Fixing typo and moving specs to correct place.
