+++ spec/controllers/admin/towns_controller_spec.rb
@@ -1,101 +1,209 @@
 require 'spec_helper'
 
 describe Admin::TownsController do
-  let(:user) { User.create!(name: 'hello', admin: true, email: 'lol@biz.info', password: 'irrelevant') }
-  let(:town) { create(:town) }
-
+  let(:user) { create :user, name: 'hello', admin: true, email: 'lol@biz.info' }
   before :each do
     sign_in user
   end
 
-  describe 'GET' do
-    context 'index' do
-      it "render template" do
-        get :index
+  describe "#update" do
+    let(:town){ Town.new(id: 123) }
+    before{
+      Town.stub(:find).and_return(town)
+      town.stub(id: 123)
+    }
 
-        expect(response).to render_template(:index)
-        expect(response).to be_success
+    let(:params){ { id: 123, town: {name: 'new town'} } }
+
+    context "that works" do
+      before{
+        Town.any_instance.stub(update_attributes: true)
+      }
+
+      it "redirects to the show path" do
+        patch :update, params
+        response.should redirect_to(admin_town_path(town))
+      end
+
+      it "responds to html" do
+        patch :update, params.merge(format: :html)
+        expect(response.content_type).to eq('text/html')
+      end
+
+      it "responds to json" do
+        patch :update, params.merge(format: :json)
+        expect(response.content_type).to eq('application/json')
       end
     end
 
-    context 'new' do
-      it "render template" do
-        get :new
+    context "that doesn't work" do
+      before{
+        Town.any_instance.stub(update_attributes: false)
+      }
 
-        expect(response).to render_template(:new)
-        expect(response).to be_success
+      context "a html request" do
+        before{ params[:format] = :html }
+
+        it "rerenders the edit path" do
+          patch :update, params
+          response.should render_template(:edit)
+        end
+
+        it "responds with html" do
+          patch :update, params.merge(format: :html)
+          expect(response.content_type).to eq('text/html')
+        end
+      end
+
+      context "a json request" do
+        before{ params[:format] = :json }
+        it "responds to json" do
+          patch :update, params.merge(format: :json)
+          expect(response.content_type).to eq('application/json')
+        end
       end
     end
 
-    context 'show' do
-      it "render template" do
-        get :show, id: town.id
+  end
 
-        expect(response).to render_template(:show)
-        expect(response).to be_success
+  describe "#create" do
+    let(:params){ { town: {name: 'new town'} } }
+
+    context "that saves ok" do
+      it "creates a town" do
+        expect{
+          post :create, params
+        }.to change { Town.count }.by(1)
+      end
+
+      it "redirects to the show path" do
+        post :create, params
+        response.should redirect_to(admin_town_path(assigns(:town)))
+      end
+
+      it "responds to html" do
+        post :create, params.merge(format: :html)
+        expect(response.content_type).to eq('text/html')
+      end
+
+      it "responds to json" do
+        post :create, params.merge(format: :json)
+        expect(response.content_type).to eq('application/json')
       end
     end
+    context "that doesn't save ok" do
+      before{ Town.any_instance.stub(save: false) }
 
-    context 'edit' do
-      it "render template" do
-        get :edit, id: town.id
+      it "does not create a town" do
+        expect{
+          post :create, params
+        }.to_not change { Town.count }
+      end
 
-        expect(response).to render_template(:edit)
-        expect(response).to be_success
+      it "rerenders the new template" do
+        post :create, params
+        response.should render_template(:new)
+      end
+
+      it "responds to html" do
+        post :create, params.merge(format: :html)
+        expect(response.content_type).to eq('text/html')
+      end
+
+      describe "a json request" do
+        it "responds with json" do
+          post :create, params.merge(format: :json)
+          expect(response.content_type).to eq('application/json')
+        end
+
+        it "returns unprocessable_entity status" do
+          post :create, params.merge(format: :json)
+          expect(response.status).to eq(422)
+        end
       end
     end
   end
 
-  describe 'POST' do
-    context 'create' do
-      it "new town" do
-        post :create, town: { name: 'Brno' }
-
-        expect(response).to redirect_to admin_town_path(assigns(:town))
-      end
+  describe "#index" do
+    it "assigns all towns to @towns" do
+      get :index
+      expect(assigns[:towns]).to eq(Town.all)
     end
 
-    context 'create json' do
-      it "new town" do
-        post :create, town: { name: 'Brno' }, format: :json
-
-        expect(response.body).to eql(assigns(:town).to_json)
-      end
-    end
-  end
-
-  describe 'PUT' do
-    context 'update' do
-      it "change name of the town" do
-        put :update, id: town.id, town: { name: 'Brno1' }
-
-        expect(response).to redirect_to admin_town_path(town)
-      end
+    it "responds to html" do
+      get :index, format: :html
+      expect(response.content_type).to eq('text/html')
     end
 
-    context 'updatea json' do
-      it "change name of the town" do
-        put :update, id: town.id, town: { name: 'Brno1' }, format: :json
-        expect(town.reload.name).to eql('Brno1')
-      end
+    it "responds to json" do
+      get :index, format: :json
+      expect(response.content_type).to eq('application/json')
     end
   end
 
-  describe 'DELETE' do
-    context 'destroy' do
-      it "change name of the town" do
-        delete :destroy, id: town.id
+  describe "#show" do
+    let(:mock_town){ Town.new(id: 123, name: 'mock town') }
+    before{
+      Town.stub(:find).and_return(mock_town)
+    }
 
-        expect(response).to redirect_to admin_towns_url
-      end
+    it "gets the right town" do
+      Town.should_receive(:find).with('123').and_return(mock_town)
+      get :show, id: 123
     end
 
-    context 'updatea json' do
-      it "change name of the town" do
-        delete :destroy, id: town.id, format: :json
-
-        expect(Town.count).to eql(0)
-      end
+    it "assigns the town" do
+      get :show, id: 123
+      expect(assigns[:town]).to eq(mock_town)
     end
+
+    it "responds to html" do
+      get :show, id: 123, format: :html
+      expect(response.content_type).to eq('text/html')
+    end
+
+    it "responds to json" do
+      get :show, id: 123, format: :json
+      expect(response.content_type).to eq('application/json')
+    end
+  end
+
+  describe "#new" do
+
+    it "assigns a new town" do
+      get :new
+      expect(assigns[:town]).to be_a(Town)
+    end
+
+
+    it "responds to html" do
+      get :new, format: :html
+      expect(response.content_type).to eq('text/html')
+    end
+
+    it "responds to json" do
+      get :new, format: :json
+      expect(response.content_type).to eq('application/json')
+    end
+  end
+
+  describe "#edit" do
+    let(:mock_town){ Town.new(id: 123, name: 'mock town') }
+    before{
+      Town.stub(:find).and_return(mock_town)
+    }
+
+    it "gets the right town" do
+      Town.should_receive(:find).with('123').and_return(mock_town)
+      get :edit, id: 123
+    end
+  end
+
+  it "remove town on destroy" do
+    at = Town.create!
+    expect {
+      post :destroy, id: at.id
+      response.should redirect_to(admin_towns_path)
+    }.to change { Town.count }.by(-1)
   end
 end
